+++ src/Window.cpp
@@ -39,7 +39,7 @@ void ObscureEngine::WS::Window::makeContextCurrent()
   this->context_makeContextCurrent();
 }
 
-void ObscureEngine::WS::Window::clear(uint8_t red, uint8_t green, uint8_t blue, uint8_t alpha)
+void ObscureEngine::WS::Window::clear(float red, float green, float blue, float alpha)
 {
   this->context_swapBuffers();
   this->context_clear(red, green, blue, alpha);
